+++ mrpostgres/sequence/generator_options.go
@@ -7,7 +7,7 @@ type (
 
 // WithMaxIDsOneQuery - устанавливает максимально возможное
 // получение ID из последовательности за один запрос к БД.
-func WithMaxIDsOneQuery(value uint64) Option {
+func WithMaxIDsOneQuery(value int) Option {
 	return func(g *Generator) {
 		g.maxIDsOneQuery = value
 	}
