+++ mrstorage/sequence.go
@@ -6,6 +6,6 @@ type (
 	// SequenceGenerator - генерирует последовательность из натуральных чисел.
 	SequenceGenerator interface {
 		Next(ctx context.Context) (nextID uint64, err error)
-		MultiNext(ctx context.Context, count uint64) (nextIDs []uint64, err error)
+		MultiNext(ctx context.Context, count int) (nextIDs []uint64, err error)
 	}
 )
